85% terminado.

+++ src/main/java/repository/InscripcionRepository.java
@@ -4,6 +4,6 @@ import entity.Inscripcion;
 import org.springframework.stereotype.Repository;
 
 @Repository("InscripcionRepository")
-public interface InscripcionRepository extends BaseRepository<Inscripcion, Integer> {
-    //todo inscripcion repo
+public interface InscripcionRepository extends BaseRepository<Inscripcion, Long> {
+    //esto puede quedar vacio.
 }
